requirements.txt
aws-cdk-lib==2.3.0
constructs>=10.0.0,<11.0.0
pylint>=2.12.2
